+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/sketch/climbing/gradient/Nadam.java
@@ -42,7 +42,7 @@ public final class Nadam extends AbstractClimber {
   public Nadam(Config config) {
     var settings = new NadamSettings(config);
     int maximumSize = Math.toIntExact(settings.maximumSize());
-    sampleSize = (int) (settings.percentSample() * maximumSize);
+    sampleSize = (long) (settings.percentSample() * maximumSize);
     stepSize = (int) (settings.percentPivot() * maximumSize);
     epsilon = settings.epsilon();
     beta1 = settings.beta1();
