+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/sketch/climbing/hill/SimpleClimber.java
@@ -28,7 +28,7 @@ public final class SimpleClimber extends AbstractClimber {
   private final double restartThreshold;
   private final double initialStepSize;
   private final double sampleDecayRate;
-  private final int initialSampleSize;
+  private final long initialSampleSize;
   private final double stepDecayRate;
   private final double tolerance;
 
@@ -37,8 +37,8 @@ public final class SimpleClimber extends AbstractClimber {
 
   public SimpleClimber(Config config) {
     var settings = new SimpleClimberSettings(config);
-    int maximumSize = Math.toIntExact(settings.maximumSize());
-    this.initialSampleSize = (int) (settings.percentSample() * maximumSize);
+    long maximumSize = settings.maximumSize();
+    this.initialSampleSize = (long) (settings.percentSample() * maximumSize);
     this.initialStepSize = settings.percentPivot() * maximumSize;
     this.restartThreshold = settings.restartThreshold();
     this.sampleDecayRate = settings.sampleDecayRate();
@@ -65,9 +65,9 @@ public final class SimpleClimber extends AbstractClimber {
     super.resetSample(hitRate);
 
     stepSize *= stepDecayRate;
-    sampleSize = (int) (sampleSize * sampleDecayRate);
+    sampleSize = (long) (sampleSize * sampleDecayRate);
     if ((stepSize <= 0.01) || (sampleSize <= 1)) {
-      sampleSize = Integer.MAX_VALUE;
+      sampleSize = Long.MAX_VALUE;
     }
   }
